+++ clients/algoliasearch-client-dart/packages/algoliasearch/lib/src/model/recommendations_results.g.dart
@@ -68,11 +68,11 @@ RecommendationsResults _$RecommendationsResultsFromJson(
           queryID: $checkedConvert('queryID', (v) => v as String?),
           automaticInsights:
               $checkedConvert('_automaticInsights', (v) => v as bool?),
-          page: $checkedConvert('page', (v) => (v as num).toInt()),
-          nbHits: $checkedConvert('nbHits', (v) => (v as num).toInt()),
-          nbPages: $checkedConvert('nbPages', (v) => (v as num).toInt()),
+          page: $checkedConvert('page', (v) => (v as num?)?.toInt()),
+          nbHits: $checkedConvert('nbHits', (v) => (v as num?)?.toInt()),
+          nbPages: $checkedConvert('nbPages', (v) => (v as num?)?.toInt()),
           hitsPerPage:
-              $checkedConvert('hitsPerPage', (v) => (v as num).toInt()),
+              $checkedConvert('hitsPerPage', (v) => (v as num?)?.toInt()),
           hits: $checkedConvert('hits', (v) => v as List<dynamic>),
         );
         return val;
@@ -119,10 +119,10 @@ Map<String, dynamic> _$RecommendationsResultsToJson(
   writeNotNull('userData', instance.userData);
   writeNotNull('queryID', instance.queryID);
   writeNotNull('_automaticInsights', instance.automaticInsights);
-  val['page'] = instance.page;
-  val['nbHits'] = instance.nbHits;
-  val['nbPages'] = instance.nbPages;
-  val['hitsPerPage'] = instance.hitsPerPage;
+  writeNotNull('page', instance.page);
+  writeNotNull('nbHits', instance.nbHits);
+  writeNotNull('nbPages', instance.nbPages);
+  writeNotNull('hitsPerPage', instance.hitsPerPage);
   val['hits'] = instance.hits.toList();
   return val;
 }