+++ clients/algoliasearch-client-dart/packages/client_recommend/lib/src/model/search_pagination.g.dart
@@ -12,20 +12,28 @@ SearchPagination _$SearchPaginationFromJson(Map<String, dynamic> json) =>
       json,
       ($checkedConvert) {
         final val = SearchPagination(
-          page: $checkedConvert('page', (v) => (v as num).toInt()),
-          nbHits: $checkedConvert('nbHits', (v) => (v as num).toInt()),
-          nbPages: $checkedConvert('nbPages', (v) => (v as num).toInt()),
+          page: $checkedConvert('page', (v) => (v as num?)?.toInt()),
+          nbHits: $checkedConvert('nbHits', (v) => (v as num?)?.toInt()),
+          nbPages: $checkedConvert('nbPages', (v) => (v as num?)?.toInt()),
           hitsPerPage:
-              $checkedConvert('hitsPerPage', (v) => (v as num).toInt()),
+              $checkedConvert('hitsPerPage', (v) => (v as num?)?.toInt()),
         );
         return val;
       },
     );
 
-Map<String, dynamic> _$SearchPaginationToJson(SearchPagination instance) =>
-    <String, dynamic>{
-      'page': instance.page,
-      'nbHits': instance.nbHits,
-      'nbPages': instance.nbPages,
-      'hitsPerPage': instance.hitsPerPage,
-    };
+Map<String, dynamic> _$SearchPaginationToJson(SearchPagination instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('page', instance.page);
+  writeNotNull('nbHits', instance.nbHits);
+  writeNotNull('nbPages', instance.nbPages);
+  writeNotNull('hitsPerPage', instance.hitsPerPage);
+  return val;
+}